Pull out and improve agent io utils
